kmir/**: inline Tools directly into KMIR class
