+++ kmir/src/kmir/testing/__init__.py
@@ -1,1 +1,1 @@
-from .fixtures import tools
+
